Properly set dtype of operands, start work on signed/unsigned ops
First, we need to properly set the datatype of immediates & similar
operands in the analysis phase, so that later in the output phase we
can make output of the correct width; 64-bit, 32-bit, etc. This is
most obvious in immediate load instructions.

Second, we start work on recognizing immediate, phrase and displacement
operands as signed. This isn't as clearly documented as other parts
of eBPF, but it appears that these are all signed. We may need to add
handling of operand width to properly start outputting correct values.