Add bpf helper translation basics to processor
We use snippets of kernel source to (after some manual work) generate
a python array of strings of bpf helper names, in order of id, so
we can map the id used in a bpf_call instruction to the helper function
name. run gcc -E on the header, and use vim to search/replace your way
from the enum definition to a list of helper name strings. This can
be automated later, but should be periodically redone to support new
helpers.

This may need to be migrated elsewhere later if it is better done via
FLIRT, an Ida script, or some other way. Note that at some point we'll
probably also need to differentiate between bpf helpers being called,
and other bpf programs being tail-called.