Change displacement immediate width to 16 bits, disable stack vars
displacement offsets in instructions such as `stxw[r10 - 8], r1` are
16 bits, not 32 bits. This is important because they are signed. This
fixes broken output like `stxw[r10 - 0xFFF8], r1`. However there are
some load instructions like `ldinddw` which might need to be handled
as 32-bit immediates, so test this later.

This fixes disassembly for a store sequence we've come across for
placing a string on the stack, but unfortunately breaks stackvar
creation, so we've disabled that part. The stack vars are still
created with respect to a 32-bit address based on the offset, and
so are incorrectly named things like `arg_FFD8`. This also means that
when we change the operands to stack variables, they're considered
broken and highlighted in red. So, we've disabled all this for now.

Also there's misc. debug output changes.