Use ctypes to force treating jump offsets as signed 16-bit ints
We were properly ingesting 2 bytes for jump offsets, but were treating
the values as unsigned which is incorrect. In fact, eBPF "recently" does
support backwards jumps, so we should expect to see legitimate jumps with
negative offsets.

This was originally only encountered in the unconditional jump case,
but appears to also be possible for conditional jumps. If this winds up being
incorrect or causing problems later we can undo this for conditional jumps.