Add bpf helper call annotation script
This script (run via File > Script File ...) scans all instruction
bytes of all known functions for eBPF call instructions. For each
call instruction the parameter (number of eBPF helper) is looked up
in a hardcoded list of helpers. If found, the helper's name is added
as a comment for that instruction. If not found, the script likely needs
to be updated with a helper list from the targeted kernel version.
Currently this is manual.

Note that this only examines known functions, not all instructions. So you
may need to manually create functions with the P hotkey.

Currently this only emits the helper name, but the next planned
improvement is to also comment with the signature.