Add handling for lock/atomic (0xdb) instructions
This is a bit complex because the actual operation is
encoded in the immediate, not the opcode. So, our handling
doesn't fit in cleanly with the existing code, and needed
some extra work and special handling.

We still have other lock operations we haven't seen or tested
against yet, so we're leaving debug print statements in place
for anyone who may run across them.

Also the linux kernel source we're using as a reference is
AT&T-like syntax while we're generating Intel-like syntax,
so we probably will revise/refine our output in the future.