Fix disassembly of legacy cBPF absolute/indirect skb loads
These implicitly operate on a socket buffer, and appear to be
legacy cBPF instructions that are supported by eBPF but in modern
kernels are rewritten as a sequence of eBPF instructions.

Aboslute loads (BPF_ABS) use r0 as hardcoded destination, and the
immediate as the offset into an implicit skb buffer. Indirect
loads (BPF_IND) additionally use a source register as an additional
offset into the implied skb.

We leaned heavily on the linux kernel's bpf disassembler for
proper disassembly of these instructions.