Clean up helper annotation script, add scripts to generate helper lookup dicts
The helper lookup in the bpf helper annotation script is
streamlined to just be signatures, and to directly use the
dictionary mapping integer helper IDs to helper signatures.

The helper signature lookup dict was generated with the scripts
under the helper_annotation directory. Just cd to that directory,
then run ./generate_helper_lookup.sh to generate the dict
definition. This makes updating for newer kernel versions much easier.

To update the annotation script for a newer kernel version,
just set the KERNEL_VERSION environment variable to the desired tag, such
as 'v5.14' (or 'v5.11' for older), run the script, and manually
edit the annotation script to use the dictionary you just created.