Add script to annotate map relocations
This is essential for reversing. The actual address/id of a map isn't known
until load time, so it must be dynamically added to programs. libbpf does this
as part of its opening/loading process, where it creates the maps, and then
uses relocations to fix up references to these maps (as far as I've skimmed).

Based on observations of eBPF ELFs, and some skimming of libbpf source and its
quite complex loading process, we use the elf section headers and symbol/string
tables, along with program and relocation sections, to add repeatable comments
on map definitions, as well as data xrefs for relocations of these maps. The
end result is that each relocated location which references a map will have a
comment naming the map referenced, making reversing much easier. Now we can
know what map is being referenced, when before we had no idea without using
tools like llvm-objdump to show us the relocation info outside IDA.

There are many assumptions going into this script that may fail on different
inputs, so if you encounter issues compare your object using llvm-objdump with
the script's source to see what is different.

Notably, we try to replicate IDA's apparently default loading procedure for
ELFs without segments, but with PROGBIT sections. Which seems to be allocating
these sections into memory starting at 0, obeying alignment, and not allowing
sections to overlap (eg: 0-size sections). This is all just so that we can
replicate the addresses necessary for the add_dref call, so that we can let IDA
know that a particular instruction is referencing a map definition. In the
future this should probably be done by a purpose-built eBPF ELF loader, so that
we are in full control of section mapping and can guarantee that the endpoints
of this data xref are correct. Needless to say, if you have an input which winds
up with a different loading arrangement once loaded into IDA, this script will
give incorrect data xrefs, map definition comments, etc.

Use alongside the bpf helper annotation!