Include NOBITS sections as program sections when annotating maps
This was known to be an issue, just getting around to fixing it once we're
absolutely certain. This was breaking our address calculations for ELFs which
had a NOBITS section.

Where PROGBITS sections require space to be allocated by the loader in the
program's memory, and supply the bytes to put in that memory, NOBITS sections
supply no bytes. So instead, the loader allocates 0-initialized bytes in memory
based on the section's size & allocation needs.

This makes more sense if you think about having lots of 0-initialized
variables/data, but see that having all these 0x00 bytes present in your ELF
file is inefficient. Just tell the loader to do it.