Add rudimentary global annotating, rename relocation annotation script
We can nearly do for globals what we did for maps, annotate the locations
they're used using relocations. However we're a bit more limited.
We seem to have false positives, and can't always easily determine
where a global actually resides.

It's possible for globals to appear as externs in the ELF file, meaning
we don't really have a location to place a repeatable comment on, like
for maps. We can try additionally replicating how IDA creates an extern
segment for these kinds of symbols, but that is getting into territory
where I'd rather write a loader and fully control the situation than
try to replicate it.

When a global does have a defined location, we add a xref and a repeatable
comment, exactly like with maps. If it doesn't, we just directly put a
comment on the relocated instruction with the name of the global. This
at least helps read code, but unfortunately doesn't give you xrefs.